+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "fitapp",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "frontend/index.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -9,11 +9,18 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^7.3.10",
+    "@react-navigation/native": "^7.1.6",
     "expo": "^52.0.46",
     "expo-cli": "^6.3.10",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
-    "react-native": "0.76.9"
+    "react-native": "0.76.9",
+    "react-native-gesture-handler": "^2.25.0",
+    "react-native-reanimated": "^3.17.5",
+    "react-native-safe-area-context": "^5.4.0",
+    "react-native-screens": "^4.10.0",
+    "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
